Moved components connected to the store to the containers directory, fixed pathing
