+++ src/containers/color_generator/ColorGenerator.test.js
