+++ src/containers/color_generator/_color_generator.scss
