+++ src/containers/color_generator/index.js
